+++ src/_includes/blocks/text/fields.ts
@@ -7,6 +7,11 @@ export default {
     "title: text",
     "id: text",
     theme,
+    {
+      name: "columns",
+      type: "checkbox",
+      description: "Display text in columns instead a single centered column",
+    },
     "text: markdown",
   ],
 };