+++ web/_cms.ts
@@ -1,0 +1,3 @@
+import cms from "top/_cms.ts";
+
+export default cms;