+++ CHANGELOG.md
@@ -4,6 +4,15 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/) and this
 project adheres to [Semantic Versioning](https://semver.org/).
 
+## [0.1.3] - 2024-08-03
+### Added
+- Styles for code.
+- Allow to configure the text block to be multicolumn or single column.
+- Additional block themes "Secondary box" and "Secondary full width".
+
+### Fixed
+- Improved some default styles.
+
 ## [0.1.2] - 2024-07-18
 ### Fixed
 - Remove files configuration.
@@ -16,6 +25,7 @@ project adheres to [Semantic Versioning](https://semver.org/).
 ## [0.1.0] - 2024-07-17
 First version
 
+[0.1.3]: https://github.com/tarugoconf/TOP/compare/v0.1.2...v0.1.3
 [0.1.2]: https://github.com/tarugoconf/TOP/compare/v0.1.1...v0.1.2
 [0.1.1]: https://github.com/tarugoconf/TOP/compare/v0.1.0...v0.1.1
 [0.1.0]: https://github.com/tarugoconf/TOP/releases/tag/v0.1.0